requirements.txt
kaleido
numpy
pandas
pillow
plotly